% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_egger_rjags.R
\name{mr_egger_rjags}
\alias{mr_egger_rjags}
\title{Bayesian implementation of the MR-Egger model with choice of prior distributions using JAGS.}
\usage{
mr_egger_rjags(object, methods = "default", betaprior = "",
  sigmaprior = "", n.chains = 1, n.burn = 1000, n.iter = 5000,
  seed = NULL, ...)
}
\arguments{
\item{object}{The data frame converted into the format_mr format}

\item{methods}{The values are used to identify the proposed priors default indicates non-informative prior weak indicates weakly informative prior pseudo indicates pseudo-horseshoe prior.}

\item{betaprior}{This is an option for setting a prior for the causal estimate.}

\item{sigmaprior}{This is an option for setting a prior for the inflating parameter in the MR-Egger model.}

\item{n.chains}{This is an option for choosing the number of chains for MCMC simulation, default number is 3 chains.}

\item{n.burn}{This is the option for the burn in period of the bayesian MCMC runs. The default option is 1000 samples}

\item{n.iter}{This is the option for the number of bayesian MCMC runs. The default is 5000 iterations}

\item{seed}{This is an option for setting seeds for reproducible results. The default is NULL}

\item{...}{Passing options through to rjags::jags.model()}
}
\value{
The result object of class eggerjags contains the following components:
\describe{
\item{AvgPleio}{The mean of the generated pleiotropic effect}
\item{CausalEffect}{The mean of the generated causal effects}
\item{StandardError}{Standard deviation of the mean causal effect}
\item{psi}{The value of the inflating parameter based on the priors}
\item{CredibleInterval}{The credible interval for the causal effect, which indicates the lower(2.5\%), median (50\%) and upper intervals (97.5\%)}
\item{samples}{Output of the bayesian MCMC samples with the different chains}
\item{method}{The specified prior}
}
}
\description{
Bayesian implementation of the MR-Egger model with choice of prior distributions using JAGS.
}
\examples{
\donttest{
fit <- mr_egger_rjags(bmi_insulin, n.chains = 3)
summary(fit)
plot(fit$samples)
}

}
\references{
Bowden et. al. , Mendelian randomization with invalid instruments: effect estimation and bias detection through Egger regression. International Journal of Epidemiology 2015. 44(2): p. 512-525.
}
\author{
Okezie Uche-Ikonne; Tom Palmer
}
