% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_egger_stan.R
\name{mr_egger_stan}
\alias{mr_egger_stan}
\title{Bayesian inverse variance weighted model with a choice of prior distributions fitted using RSTAN.}
\usage{
mr_egger_stan(
  data,
  prior = 1,
  n.chains = 1,
  n.burn = 200,
  n.iter = 1000,
  seed = 234,
  rho = 0.5
)
}
\arguments{
\item{data}{A data of class mr_format}

\item{prior}{An integer for selecting the prior distributions; "1" selects a non-informative set of priors; "2" selects weakly informative priors; "3" selects a pseudo-horseshoe prior on the causal effect; "4" selects joint prior of the intercept and causal effect estimate.}

\item{n.chains}{Numeric indicating the number of chains used in the HMC estimation in rstan, the default is 1 chain.}

\item{n.burn}{Numeric indicating the burn-in period of the Bayesian HMC estimation. The default is 200 samples.}

\item{n.iter}{Numeric indicating the number of iterations in the Bayesian HMC estimation. The default is 1000 iterations.}

\item{seed}{Numeric indicating the random number seed. The default is 234.}
}
\description{
Bayesian inverse variance weighted model with a choice of prior distributions fitted using RSTAN.
}
\examples{
data(bmi_insulin)
dat <- mr_format(rsid = bmi_insulin[,"rsid"],
          xbeta = bmi_insulin[,"beta.exposure"],
          ybeta = bmi_insulin[,"beta.outcome"],
          xse = bmi_insulin[,"se.exposure"],
          yse = bmi_insulin[,"se.outcome"])
egger_fit <- mr_egger_stan(dat)
print(egger_fit)
}
\references{
Burgess, S., Butterworth, A., Thompson S.G. Mendelian randomization analysis with multiple genetic variants using summarized data. Genetic Epidemiology, 2013, 37, 7, 658-665 \url{https://dx.doi.org/10.1002/gepi.21758}.
}
