% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_ivw_rjags.R
\name{mr_ivw_rjags}
\alias{mr_ivw_rjags}
\title{Fitting Bayesian inverse variance weighted estimates using established priors using the JAGS software.}
\usage{
mr_ivw_rjags(object, methods = "default", betaprior = "",
  n.chains = 1, n.burn = 1000, n.iter = 5000, seed = NULL, ...)
}
\arguments{
\item{object}{The data frame converted into the format_mr format}

\item{methods}{The values are used to identify the proposed priors default indicates non-informative prior weak indicates weakly informative prior pseudo indicates pseudo-horseshoe prior.}

\item{betaprior}{This is an option for setting a prior for the causal estimate.}

\item{n.chains}{This is an option for choosing the number of chains for MCMC simulation, default number is 3 chains.}

\item{n.burn}{This is the option for the burn in period of the bayesian MCMC runs. The default option is 1000 samples}

\item{n.iter}{This is the option for the number of bayesian MCMC runs. The default is 5000 iterations}

\item{seed}{This is an option for setting seeds for reproducible results. The default is to use the rjags default of using the current system time.}

\item{...}{Passing options through to rjags::jags.model()}
}
\value{
The result object of class ivwjags contains the following components:
\describe{
\item{CausalEffect}{The mean of the generated causal effects}
\item{StandardError}{Standard deviation of the mean causal effect}
\item{CredibleInterval}{The credible interval for the causal effect, which indicates the lower(2.5\%), median (50\%) and upper intervals (97.5\%)}
\item{samples}{Output of the bayesian MCMC samples with the different chains}
\item{method}{The specified prior}
}
}
\description{
Fitting Bayesian inverse variance weighted estimates using established priors using the JAGS software.
}
\examples{
data(bmi_insulin)
fit <- mr_ivw_rjags(bmi_insulin, n.chains = 1)
print(fit)
summary(fit)
plot(fit$samples)

}
