% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_ivw_stan.R
\name{mr_ivw_stan}
\alias{mr_ivw_stan}
\title{Bayesian inverse variance weighted model with a choice of prior distributions fitted using RSTAN.}
\usage{
mr_ivw_stan(
  data,
  prior = 1,
  n.chains = 1,
  n.burn = 200,
  n.iter = 1000,
  seed = 234
)
}
\arguments{
\item{data}{A data of class mr_format}

\item{prior}{An integer for selecting the prior distributions; "1" selects a non-informative set of priors; "2" selects weakly informative priors; "3" selects a pseudo-horseshoe prior on the causal effect.}

\item{n.chains}{Numeric indicating the number of chains used in the HMC estimation in rstan, the default is 1 chain.}

\item{n.burn}{Numeric indicating the burn-in period of the Bayesian HMC estimation. The default is 200 samples.}

\item{n.iter}{Numeric indicating the number of iterations in the Bayesian MCMC estimation. The default is 1000 iterations.}

\item{seed}{Numeric indicating the random number seed. The default is 234.}
}
\description{
Bayesian inverse variance weighted model with a choice of prior distributions fitted using RSTAN.
}
\examples{
ivw_fit <- mr_ivw_stan(bmi_insulin)
print(ivw_fit)
}
\references{
Burgess, S., Butterworth, A., Thompson S.G. Mendelian randomization analysis with multiple genetic variants using summarized data. Genetic Epidemiology, 2013, 37, 7, 658-665 \url{https://dx.doi.org/10.1002/gepi.21758}.
}
