% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_radialegger_rjags.R
\name{mr_radialegger_rjags}
\alias{mr_radialegger_rjags}
\title{Fitting Bayesian MR-radialegger models using various priors from the jags software.}
\usage{
mr_radialegger_rjags(object, prior = "default", betaprior = "",
  sigmaprior = "", n.chains = 1, n.burn = 1000, n.iter = 5000,
  seed = NULL, rho = 0.5, ...)
}
\arguments{
\item{object}{The data frame converted into the format_mr format}

\item{prior}{The option for selecting the proposed priors; "default" indicates non-informative prior; "weak" indicates weakly informative prior; "pseudo" indicates pseudo-horseshoe prior; "joint" indicates joint probability prior.}

\item{betaprior}{This is an option for setting a prior for the causal estimate.}

\item{sigmaprior}{This is an option for setting a prior for the inflating parameter in the Radial formulation of MR-Egger model.}

\item{n.chains}{This is an option for choosing the number of chains for MCMC simulation, default number is 3 chains.}

\item{n.burn}{This is the option for the burn in period of the bayesian MCMC runs. The default option is 1000 samples}

\item{n.iter}{This is the option for the number of bayesian MCMC runs. The default is 5000 iterations}

\item{seed}{This is an option for setting seeds for reproducible results. The default is NULL}

\item{rho}{The correlation coefficient factor used to analyze the joint prior method}

\item{...}{Passing options through to rjags::jags.model()}
}
\value{
The result object of class radialeggerjags contains the following components:
\describe{
\item{AvgPleio}{The mean of the generated pleiotropic effect}
\item{CausalEffect}{The mean of the generated causal effects}
\item{StandardError}{Standard deviation of the mean causal effect}
\item{sigma}{The value of the inflating parameter based on the priors}
\item{CredibleInterval}{The credible interval for the causal effect, which indicates the lower(2.5\%), median (50\%) and upper intervals (97.5\%)}
\item{samples}{Output of the bayesian MCMC samples with the different chains}
\item{Priors}{The specified priors}
}
}
\description{
Fitting Bayesian MR-radialegger models using various priors from the jags software.
}
\examples{
data(bmi_insulin)
fit <- mr_radialegger_rjags(bmi_insulin, n.chains = 1)
summary(fit)
plot(fit$samples)

}
\references{
Bowden, J., et al., Improving the visualization, interpretation and analysis of two-sample summary data Mendelian randomization via the Radial plot and Radial regression. International Journal of Epidemiology, 2018. 47(4): p. 1264-1278. \url{https://doi.org/10.1093/ije/dyy101}
}
