% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbayes-package.r
\docType{package}
\name{mrbayes-package}
\alias{mrbayes-package}
\title{mrbayes: Bayesian implementation of the IVW and MR-Egger models for two-sample Mendelian randomization analyses}
\description{
mrbayes: Bayesian implementation of the IVW and MR-Egger models for two-sample Mendelian randomization analyses
}
