% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbayes-package.r
\docType{package}
\name{mrbayes-package}
\alias{mrbayes-package}
\alias{mrbayes}
\title{mrbayes: Bayesian implementation of the IVW and MR-Egger models for two-sample Mendelian randomization analyses}
\description{
Bayesian implementation of the IVW and MR-Egger models for two-sample Mendelian randomization analyses.
}
\references{
Stan Development Team (2019). RStan: the R interface to Stan. R package version 2.19.2. \url{https://mc-stan.org}
}
